Use doubles exclusively in Rendering timers

+++ include/render.h
@@ -55,7 +55,7 @@ typedef struct {
 		unsigned bpp;
 		bool dblw,dblh;
 		double ratio;
-		float fps;
+		double fps;
 	} src;
 	struct {
 		int count;
@@ -93,8 +93,13 @@ typedef struct {
 
 extern Render_t render;
 extern ScalerLineHandler_t RENDER_DrawLine;
-void RENDER_SetSize(Bitu width, Bitu height, unsigned bpp, float fps,
-                    double ratio, bool dblw, bool dblh);
+void RENDER_SetSize(Bitu width,
+                    Bitu height,
+                    unsigned bpp,
+                    double fps,
+                    double ratio,
+                    bool dblw,
+                    bool dblh);
 bool RENDER_StartUpdate(void);
 void RENDER_EndUpdate(bool abort);
 void RENDER_SetPal(Bit8u entry,Bit8u red,Bit8u green,Bit8u blue);
